requirements.txt
# Smart-RAG-Chatbot Requirements
# Python dependencies for utility scripts and workflow management

# Core dependencies
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.3
python-dotenv>=1.0.0
requests>=2.31.0
click>=8.1.7
rich>=13.7.0
pyyaml>=6.0.1

# Vector Database
qdrant-client>=1.7.0

# AI/ML Libraries
openai>=1.12.0
google-generativeai>=0.3.2

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.12.1
flake8>=7.0.0
mypy>=1.8.0

# Utilities
tqdm>=4.66.1
colorama>=0.4.6
tabulate>=0.9.0

# Data handling
pandas>=2.1.4
